# websocket: improve frame parsing