Adjust to avoid creating a temporary buffer