@KhafraDev 2024-08-12
I'd prefer if there was a benchmark to measure the change.