@tsctx 2024-08-12
As an example, the 256Kib receive benchmark is 1.2x faster.