@KhafraDev 2024-08-13
Benchmarking websockets is hard and there is no benchmark to verify.