@tsctx 2024-10-15
@KhafraDev Would you be able to revisit this? This is the result of the #3203's of benchmarks.

before / after

```shell
> $ node ./benchmarks/websocket-benchmark.mjs
(node:11148) [UNDICI-WSS] Warning: WebSocketStream is experimental! Expect it to change at any time.
(Use `node --trace-warnings ...` to show where the warning was created)
undici [binary]: transferred 95.78MiB Bytes/s
undici [string]: transferred 92.98MiB Bytes/s
undici - stream [binary]: transferred 95.93MiB Bytes/s
undici - stream [string]: transferred 97.00MiB Bytes/s
ws [binary]: transferred 125.82MiB Bytes/s
ws [string]: transferred 117.20MiB Bytes/s
> $ node ./benchmarks/websocket-benchmark.mjs
(node:3924) [UNDICI-WSS] Warning: WebSocketStream is experimental! Expect it to change at any time.
(Use `node --trace-warnings ...` to show where the warning was created)
undici [binary]: transferred 113.48MiB Bytes/s
undici [string]: transferred 112.11MiB Bytes/s
undici - stream [binary]: transferred 106.25MiB Bytes/s
undici - stream [string]: transferred 91.41MiB Bytes/s
ws [binary]: transferred 122.96MiB Bytes/s
ws [string]: transferred 114.50MiB Bytes/s
```